Add Components Recommender mock plugin. Closes #63
