+++ src/common/plugins/README.md
@@ -62,3 +62,12 @@ In the above JSON, `id` can be one of the following:
 The rest of the fields correspond to the similarly named concepts in WebGME. For more information, check out https://webgme.readthedocs.io/en/latest/meta_modeling/meta_modeling_concepts.html (or the [source docs](https://editor.webgme.org/docs/source/Core.html) for usage with the Core).
 
 Example usage for this utility can be found in the [tests](/test/common/plugins/JSONImporter.spec.js) and in the [CreateKerasMeta](/src/plugins/CreateKerasMeta/CreateKerasMeta.js) plugin.
+
+## CircuitAnalysisBases
+The base implementations for all circuit analysis/conversion Plugins. This module implements a conversion from a Circuit(WebGME) to a Circuit(PySpice) and a base class for running analysis on the Circuit.
+
+Currently these implementations are used by the following plugins:
+
+1. [ConvertCircuitToNetlist](../../plugins/ConvertCircuitToNetlist/ConvertCircuitToNetlist/__init__.py): Converts a WebGME Circuit to its equivalent SPICE Netlist (uses CircuitToPySpiceBase)
+2. [RecommendNextComponentsMock](../../plugins/RecommendNextComponentsMock/RecommendNextComponentsMock/__init__.py): A mock implementation for recommending components to be added to the Circuit (uses AnalyzeCircuit)
+
