+++ src/plugins/RecommendNextComponentsMock/metadata.json
@@ -1,0 +1,15 @@
+{
+  "id": "RecommendNextComponentsMock",
+  "name": "RecommendNextComponentsMock",
+  "version": "0.1.0",
+  "description": "",
+  "icon": {
+    "class": "glyphicon glyphicon-cog",
+    "src": ""
+  },
+  "disableServerSideExecution": false,
+  "disableBrowserSideExecution": true,
+  "dependencies": [],
+  "writeAccessRequired": false,
+  "configStructure": []
+}
